add fee cli
