+++ crates/fiber-cli/build.rs
@@ -537,6 +537,7 @@ const GROUP_ORDER: &[&str] = &[
     "channel",
     "invoice",
     "payment",
+    "fee",
     "graph",
     "cch",
     "dev",
