+++ tests/nodes/deployer/config.yml
@@ -17,6 +17,7 @@ rpc:
     - payment
     - graph
     - info
+    - fee
     - invoice
     - peer
     - watchtower
